refactor(frontend): migrate graph detail dialogs to shared Dialog shell
